mach-snapdragon: move memory parsing to its own file
This code is getting a bit complicated, split it out to try and keep
things a bit better organised as we're going to be supporting populating
the memory layout from various other sources.

Tested-by: Sumit Garg <sumit.garg@oss.qualcomm.com>
